requirements.txt
pycollada
numpy
